move match_histogram to expsure
