+++ TODO.txt
@@ -37,6 +37,7 @@ Version 0.18
 * In ``skimage/morphology/convex_hull.py`` remove the deprecated parameter
   ``neighbors`` and update the tests. Set the default value of ``connectivity``
   to 2.
+* In ``skimage/transform`` remove histogram_matching.py.
 
 Other
 -----
